Slot-based composition model with core/convention separation
Schema v3: slots, entry_slots, deploy_root, compose_mode
Compose: convention-agnostic slot iterator
Scanner: ComposeMode from conventions, multi-extension, singleton exclusion
Bootstrap: deploy_root per consumer, slot generation, container consumer
Registry: Close() on interface, New() factory, slot CRUD, deployment_trace
Deployer: EnsureGitExclude for non-opt-in repos
Query: ListSlots/ListSlotEntries
CLI: run updated for slot-based compose, plan uses slots
OS configs: ported from carula to carrel/omp/config/, correct OS dest paths
Arch-design: updated for slot model and core/convention separation
Rules: beads discipline (never close without verify)

+++ internal/authoring/authoring.go
@@ -36,7 +36,7 @@ func AddEntry(reg registry.Registry, typ registry.CapabilityType, name string, s
 		Name:         name,
 		Type:         typ,
 		RelativePath: relPath,
-		ContentHash:  xxhash.Sum64(content),
+		ContentHash:  int64(xxhash.Sum64(content)),
 		CreatedBy:    registry.OriginCarrel,
 	}
 
@@ -78,7 +78,7 @@ func EditEntry(reg registry.Registry, typ registry.CapabilityType, name string, 
 		return err
 	}
 
-	newHash := xxhash.Sum64(newContent)
+	newHash := int64(xxhash.Sum64(newContent))
 	entry.ContentHash = newHash
 
 	// Re-register to update the stored entry.
